refactor(discord): #4229 S5 — extract post stream exit (move-pure)
Verbatim-move the post-stream-exit finalize tail (225 lines, L4989-5213)
of `tmux_output_watcher_with_restore` out of the tmux_watcher.rs giant
into a new child module tmux_watcher/post_stream_exit.rs.

The moved tail performs: DashMap watcher-slot removal (#243-safe,
cancel-gated), dispatch-protection resolve + dead-session dispatch fail,
pane-dead explicit inflight cleanup (spawn_blocking revalidate), dead
tmux session kill + post-mortem capture (unified-thread auto-queue
skip), idle-status report, and the watcher-stopped log.

Behavior-preserving decompose:
- Block bytes are char-identical to the original (verified by diff of
  the pre-move extraction against the pasted block).
- fn-level return/break/continue mappings = 0 (the two `return;` sit
  inside spawn_blocking `move ||` closures and move verbatim).
- 6 threaded locals passed via PostStreamExitContext (owned move; the
  call is the fn tail so no post-move use conflict): channel_id, shared,
  tmux_session_name, cancel, watcher_turn_identity, watcher_instance_id.
- No macro_rules / const hoist; no direct Discord send in the block
  (audit_allowlist unchanged). New child uses `use super::*` + explicit
  `use std::sync::atomic::{AtomicBool, Ordering}`.

tmux_watcher.rs 5218 -> 5007 raw/prod (-211). Ratchet + giant baseline
lowered to 5007; inventory docs regenerated. New child (254 LoC) stays
under the tmux_watcher/** 700-cap so no admission/registry entry needed.
