+++ src/services/discord/tmux_watcher/post_stream_exit.rs
@@ -1,0 +1,254 @@
+use super::*;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
+
+/// #4229 S5: post-stream-exit finalize tail of `tmux_output_watcher_with_restore`
+/// (DashMap slot removal, dispatch-protection resolve + dead-session dispatch fail,
+/// pane-dead explicit inflight cleanup, dead tmux session kill + post-mortem capture,
+/// idle-status report, watcher-stopped log), moved verbatim from tmux_watcher.rs.
+pub(super) struct PostStreamExitContext {
+    pub(super) channel_id: ChannelId,
+    pub(super) shared: Arc<SharedData>,
+    pub(super) tmux_session_name: String,
+    pub(super) cancel: Arc<AtomicBool>,
+    pub(super) watcher_turn_identity:
+        Option<crate::services::discord::inflight::InflightTurnIdentity>,
+    pub(super) watcher_instance_id: u64,
+}
+
+pub(super) async fn run_post_stream_exit(ctx: PostStreamExitContext) {
+    let PostStreamExitContext {
+        channel_id,
+        shared,
+        tmux_session_name,
+        cancel,
+        watcher_turn_identity,
+        watcher_instance_id,
+    } = ctx;
+
+    // Cleanup: only remove from DashMap if we weren't cancelled/replaced.
+    // #243: When a watcher is cancelled (replaced by a new watcher or shutdown),
+    // the replacement already occupies the slot — removing would delete the new entry.
+    if !cancel.load(Ordering::Relaxed) {
+        shared.tmux_watchers.remove(&channel_id);
+    }
+
+    let api_port = shared.api_port;
+    let provider = shared.settings.read().await.provider.clone();
+    let session_key = crate::services::discord::adk_session::build_adk_session_key(
+        &shared, channel_id, &provider,
+    )
+    .await;
+    let channel_name = {
+        let data = shared.core.lock().await;
+        data.sessions
+            .get(&channel_id)
+            .and_then(|s| s.channel_name.clone())
+    };
+    let dispatch_protection =
+        crate::services::discord::tmux_lifecycle::resolve_dispatch_tmux_protection(
+            shared.pg_pool.as_ref(),
+            &shared.token_hash,
+            &provider,
+            &tmux_session_name,
+            channel_name.as_deref(),
+        );
+    let dispatch_failed_for_dead_session = if let Some(protection) = dispatch_protection.as_ref() {
+        crate::services::discord::tmux_lifecycle::fail_active_dispatch_for_dead_tmux_session(
+            api_port,
+            protection,
+            &tmux_session_name,
+            "tmux_watcher",
+        )
+        .await
+    } else {
+        false
+    };
+    let cleanup_plan = dead_session_cleanup_plan(
+        dispatch_protection.is_some() && !dispatch_failed_for_dead_session,
+    );
+
+    if let Some(protection) = dispatch_protection {
+        let ts = chrono::Local::now().format("%H:%M:%S");
+        if dispatch_failed_for_dead_session {
+            tracing::warn!(
+                "  [{ts}] tmux watcher: failed active dispatch for dead session {} — {}",
+                tmux_session_name,
+                protection.log_reason()
+            );
+        } else {
+            tracing::info!(
+                "  [{ts}] ♻ tmux watcher: preserving dispatch session {} — {}",
+                tmux_session_name,
+                protection.log_reason()
+            );
+        }
+    }
+
+    if !cleanup_plan.preserve_tmux_session {
+        // #2427 A wire: pane-death explicit inflight cleanup. The
+        // tmux pane is gone (or about to be killed below), so any
+        // inflight row still pointing at this provider/channel will
+        // never receive a normal completion hook. Without this the
+        // sweeper has to time-guess (`STALL`/`ABANDON`) before evicting,
+        // reproducing the #2415 family of "completion-missing → time
+        // heuristic" bugs.
+        //
+        // We re-check `tmux_session_has_live_pane` on the blocking
+        // thread before clearing, matching the same revalidation the
+        // kill path uses (#1261 codex P2) so a concurrent
+        // `start_claude` respawn of a fresh same-named session does not
+        // get its inflight wiped.
+        {
+            let sess_for_inflight = tmux_session_name.clone();
+            let provider_for_inflight = provider.clone();
+            let channel_id_inflight = channel_id;
+            let watcher_identity_for_inflight = watcher_turn_identity.clone();
+            let _ = tokio::task::spawn_blocking(move || {
+                let pane_alive = tmux_session_has_live_pane(&sess_for_inflight);
+                if pane_alive {
+                    // Pane resurrected (e.g. start_claude respawn race) —
+                    // do not touch its inflight.
+                    return;
+                }
+                emit_explicit_inflight_cleanup_signal_pane_dead(
+                    &provider_for_inflight,
+                    channel_id_inflight,
+                    &sess_for_inflight,
+                    watcher_identity_for_inflight.as_ref(),
+                );
+            })
+            .await;
+        }
+
+        // Kill dead tmux session to prevent accumulation (especially for thread sessions
+        // which are created per-dispatch and would otherwise linger for 24h).
+        // #145: skip kill for unified-thread sessions with active auto-queue runs.
+        {
+            let sess = tmux_session_name.clone();
+            let _ = tokio::task::spawn_blocking(move || {
+                if tmux_session_exists(&sess) && !tmux_session_has_live_pane(&sess) {
+                    // Check if this is a unified-thread session before killing
+                    if let Some((_, ch_name)) =
+                        crate::services::provider::parse_provider_and_channel_from_tmux_name(&sess)
+                    {
+                        if crate::dispatch::is_unified_thread_channel_name_active(&ch_name) {
+                            return;
+                        }
+                    }
+                    crate::services::termination_audit::record_termination_for_tmux(
+                        &sess,
+                        None,
+                        "tmux_watcher",
+                        "dead_after_turn",
+                        Some("watcher cleanup: dead session after turn"),
+                        None,
+                    );
+                    record_tmux_exit_reason(&sess, "watcher cleanup: dead session after turn");
+
+                    // #1261 (Fix B): the wrapper's stderr `[stderr] ...` lines and
+                    // synthetic `[fatal startup error]` markers go to the PTY, not
+                    // to the structured jsonl that `recent_output_tail` reads. Dump
+                    // the current pane buffer to a `death_pane_log` file BEFORE we
+                    // kill the session so the wrapper-level death context is still
+                    // recoverable post-mortem. Kept out of `cleanup_session_temp_files`
+                    // EXTS on purpose — the file persists past the cleanup and is
+                    // overwritten on the next death of the same session.
+                    if let Some(pane_content) =
+                        crate::services::platform::tmux::capture_pane(&sess, -1000)
+                    {
+                        let stamped = format!(
+                            "[{}] post-mortem capture for session={}\n{}",
+                            chrono::Local::now().format("%Y-%m-%d %H:%M:%S"),
+                            sess,
+                            pane_content
+                        );
+                        let path = crate::services::tmux_common::session_temp_path(
+                            &sess,
+                            "death_pane_log",
+                        );
+                        if let Some(parent) = std::path::Path::new(&path).parent() {
+                            let _ = std::fs::create_dir_all(parent);
+                        }
+                        let _ = std::fs::write(&path, stamped);
+                    }
+
+                    // #1261 (codex P2): the `capture_pane` subprocess above
+                    // widens the gap between the outer dead-pane gate and the
+                    // kill. In that window a concurrent follow-up could run
+                    // claude.rs::start_claude, which kills the stale session
+                    // (line 1294), respawns a fresh live session with the
+                    // same name (line 1379), and we'd then kill the brand-new
+                    // session here. Revalidate the dead-pane condition right
+                    // before the kill so we only tear down the same
+                    // dead-paned session we capture-paned.
+                    if tmux_session_exists(&sess) && !tmux_session_has_live_pane(&sess) {
+                        crate::services::platform::tmux::kill_session(
+                            &sess,
+                            "watcher cleanup: dead session after turn",
+                        );
+                    }
+                    // NOTE: jsonl/FIFO/etc. cleanup intentionally NOT done here.
+                    // `claude.rs::start_claude` calls
+                    // `cleanup_session_temp_files` at spawn time
+                    // (`claude.rs:1304`) before recreating the canonical paths,
+                    // which already covers the "next-spawn against stale jsonl"
+                    // case. Pairing a watcher-side cleanup with the kill races
+                    // with that spawn-side cleanup + recreate (#1261 codex P1):
+                    // if the next message lands between our `kill_session` and
+                    // our cleanup, claude's spawn already laid down fresh files
+                    // and our cleanup deletes them, breaking the new turn.
+                    // Keep cleanup as a single-source-of-truth on the spawn
+                    // path.
+                }
+            })
+            .await;
+        }
+    }
+
+    let defer_idle_status_to_bridge =
+        crate::services::discord::inflight::load_inflight_state(&provider, channel_id.get())
+            .as_ref()
+            .is_some_and(|state| {
+                state.tmux_session_name.as_deref() == Some(tmux_session_name.as_str())
+            });
+
+    if cleanup_plan.report_idle_status && !defer_idle_status_to_bridge {
+        // Report idle status to DB so the dashboard doesn't show stale "working" state.
+        // Always report idle when the watcher exits, even if dispatch protection
+        // keeps the dead tmux session around for the active-dispatch safety path.
+        let thread_channel_id = channel_name
+            .as_deref()
+            .and_then(crate::services::discord::adk_session::parse_thread_channel_id_from_name);
+        let agent_id = resolve_role_binding(channel_id, channel_name.as_deref())
+            .map(|binding| binding.role_id);
+        crate::services::discord::adk_session::post_adk_session_status(
+            session_key.as_deref(),
+            channel_name.as_deref(),
+            None, // model
+            "idle",
+            &provider,
+            None, // session_info
+            None, // tokens
+            None, // cwd
+            None, // dispatch_id
+            thread_channel_id,
+            Some(channel_id),
+            agent_id.as_deref(),
+            api_port,
+        )
+        .await;
+    } else if cleanup_plan.report_idle_status {
+        tracing::debug!(
+            provider = %provider.as_str(),
+            channel = channel_id.get(),
+            tmux_session = %tmux_session_name,
+            "watcher deferred idle status because bridge-owned inflight still needs terminal Discord finalization"
+        );
+    }
+
+    let ts = chrono::Local::now().format("%H:%M:%S");
+    tracing::info!(
+        "  [{ts}] 👁 tmux watcher stopped for #{tmux_session_name} (instance {watcher_instance_id})"
+    );
+}
